Working on MVC stuff.
This is MVC, but with templates in the mix to help make the views not so full of spaghetti code.  The view should contain logic, and parse that into a plain text file that has placeholders (a *template*)... because then designers don't have to be developers, they don't have to understand the logic, they just need to be able to look at the templates and understand how they work into the bigger picture.